cloud/gcp/flask_prediction_server/requirements.txt
flask
flask-cors
google-cloud-pubsub
numpy